bad way seperated
